requirements.txt
click==6.7
cryptography==1.7.1
python-dotenv==0.10.3
Flask==1.0.2
idna==2.2
itsdangerous==0.24
Jinja2==2.10.1
keyring==10.1
keyrings.alt==1.3
MarkupSafe==1.0
pyasn1==0.1.9
pycrypto==2.6.1
RPi.GPIO==0.6.3
SecretStorage==2.3.1
six==1.10.0
Werkzeug==0.14.1
